requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dotenv==1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
